j1939: bus: let j1939_name_to_sa return a u8

+++ net/can/j1939/bus.c
@@ -118,7 +118,7 @@ struct j1939_ecu *j1939_ecu_find_by_addr(u8 sa, int ifindex)
 	return ecu;
 }
 
-int j1939_name_to_sa(name_t name, int ifindex)
+u8 j1939_name_to_sa(name_t name, int ifindex)
 {
 	struct j1939_ecu *ecu;
 	struct j1939_priv *priv;
